Backend Cleanup folder structure
