+++ server/controllers/logController.js
@@ -9,43 +9,10 @@
  * ************************************
  */
 
-import LogModel from '../models/logModel.js';
+import addLog from './logController/addLog.js';
+import getLogs from './logController/getLogs.js';
 
-const logController = {};
-
-logController.getLogs = async (req, res, next) => {
-  try {
-    if (res.locals.cookieStatus) {
-      const data = await LogModel.find({});
-      res.locals.logs = data;
-    }
-
-    return next();
-  } catch (err) {
-    return next({
-      log: `logController.getLogs: ERROR - ${err}`,
-      status: err.status || 500,
-      message: {
-        err: 'Error occured in logController.getLogs. Check server logs for more details.',
-      },
-    });
-  }
+export default {
+  addLog,
+  getLogs,
 };
-
-logController.addLog = async (req, res, next) => {
-  try {
-    const params = req.body;
-    const data = await LogModel.create(params);
-    next();
-  } catch (err) {
-    return next({
-      log: `logController.addLog: ERROR - ${err}`,
-      status: err.status || 500,
-      message: {
-        err: 'Error occured in logController.addLog. Check server logs for more details.',
-      },
-    });
-  }
-}
-
-export default logController;
